Do not unreview pages when they are moved
I369af79468a0a79b69923b9d27db555dd1cf9e20 unintentionally
made PageTriage evaluate `self::addToPageTriageQueue(...)` on
almost every page. This caused two issues:

- Already reviewed pages were enmasse added to the NewPagesFeed
when they were moved.
- Pages marked as reviewed were unreviewed by moves

This patch fixes both those issues by partially reverting some
of the changes in the previous patch and makes sure that:

- The recreated update is only applied when a page is already in
the queue and the move is not a cross-namespace move
- The `self::addToPageTriageQueue(...)` function is only called
on cross-namespace moves

A test is also added to test for this behavior.

Bug: T370593
Change-Id: Ibba7118a70830862da1506b086515949dcc03983