delete: Review pages after deletion tagging them
Fixes a bug where sending a redirect to RFD did not mark the
redirect as reviewed. This is because the order was 1) mark as
reviewed, 2) apply RFD template, but applying the template
converts the redirect into an article, which unreviews it. Moving
the review code down should fix the issue.

Bug: T382996
Change-Id: Ibb4f0ad5f4a9e10ff76cb9aa397511ed9447863f