Remove unused config variables in PageTriage
PageTriagePagesPerRequest - Some kind of deprecated functionality to control how many pages were returned by the PageTriage API (this functionality was not present when
this line was added and I don't think was ever implemented).

PageTriageMarkPatrolledLinkExpiry - How long before users are unable to patrol articles, this functionality no longer exists in the extension.

PtTemplatePath - Some kind of contrived way to load backbone.js templates, this is done by ResourceLoader now and all of the backbone.js stuff is on it's way out anyway.

PageTriageTagsOptionsMessages, PageTriageDeletionTagsOptionsMessages
- These messages were obsoleted by the migration to a json file for tags and deletions, we just hardcode them in the json file now.

Bug: T393418
Change-Id: I1ad0a98b769430a2abab6dd20fe2784244f2a07b