NPF: Fix username and keyword filter bug
On NPF, when you selected the username or the keyword radio buttons
and did not add text to the accompanying input, NPF would filter
by a blank username or keyword, which would yield no results in the
query. It would also show an active filters chip with a i18n param.

By adding a guard clause, we can now detect if the keyword or username
inputs are left blank and go back to the default 'all' setting.

Bug: T394100
Change-Id: Id7b22df9e6474e33e11a7350b090762b93cedcea