tests: move maintenance classes back into maintenance folder
In previous patches, I made it so that maintenance scripts in the
PageTriage repo `require_once` classes located in
includes/Maintenance. I thought this was necessary for phpunit
testing. However I was incorrect. It is more standard to just keep
these classes in the maintenance scripts. Testing works just fine
once you add the maintenance scripts to extension.json ->
AutoloadClasses.

Change-Id: I48da2554ad2811c5136bc69c31a12318b2766585