Convert WikiLove component to Vue
This change introduces a WikiLove component to the toolbar, originally
written in Backbone. Since this is the first tool with a flyout being
added to Vue, some adjustments needed to be made to existing files such
as `App.vue`, `ToolFlyout.vue`, `ToolIcon.vue`, `extension.json`,
`ext.pageTriage.article.js`, etc.

Bug: T349387
Change-Id: I1efed8577ef6327604cb8d246d53035c19b0438d