Prevent Curation toolbar from preventDefaulting all left click pointer events
In I1efed8577ef6327604cb8d246d53035c19b0438d there was a change to
App.vue to prevent mouse clicks from advancing further into the
DOM tree (not completely sure why). This however has the side-effect
of disabling all forms of left clicking on textboxes in the toolbar.

This patch removes the preventDefault statement since it does not appear
to make a significant difference in draggability of the toolbar even
with `pagetriage_ui=wikilove` enabled and instead has caused T404405.

Bug: T404405
Change-Id: I1469389a4b0426a385fb4d2da3c5f252285f8fb5